requirements.txt
discord.py==1.7.3
finnhub_python==2.4.5
matplotlib==3.4.3
mplfinance==0.12.7a17
pandas==1.3.4
pandas_ta==0.3.14b0
